backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
pymysql==1.1.1
redis==5.0.8
pydantic==2.9.2
python-dotenv==1.0.1
requests==2.32.3
